+++ roboticeclient/__init__.py
@@ -1,6 +1,6 @@
 
 __title__ = 'Robotice Client'
-__release__ = '5'
+__release__ = '6'
 __version__ = '0.3.%s' % __release__
 __author__ = 'Michael Kuty & Ales Komarek'
 __license__ = 'Apache 2.0'