+++ roboticeclient/common/manager.py
@@ -20,23 +20,29 @@ class BaseManager(BaseClient):
     def get(self, request, id):
         return self.request(
             request,
-            '/{0}/{1}/'.format(self.SCOPE, id),
+            '/{0}/{1}'.format(self.SCOPE, id),
             'GET')
 
 
-    def create(self, request, id, data):
+    def create(self, request, data):
         return self.request(
             request,
-            '/{0}/{1}/'.format(self.SCOPE, id),
-            'POST',
+            '/{0}/'.format(self.SCOPE),
+            'PUT',
             data)
 
     def update(self, request, id, data):
         return self.request(
             request,
             '/{0}/{1}/'.format(self.SCOPE, id),
-            'PUT',
+            'POST',
             data)
+
+    def delete(self, request, id):
+        return self.request(
+            request,
+            '/{0}/{1}/'.format(self.SCOPE, id),
+            'DELETE')
 
     def __init__(self, **kwargs):
             
@@ -44,3 +50,23 @@ class BaseManager(BaseClient):
         self.port = kwargs.pop("port", os.getenv('ROBOTICE_PORT', 9753))
         self.protocol = kwargs.pop("protocol", os.getenv('ROBOTICE_PROTOCOL', "http")) 
         self.set_api()
+
+def manager_factory(base_client=BaseManager):
+
+    def initialize(self):
+        pass
+
+    def some_event(self):
+        pass
+
+    subclass_body_dict = {
+        "initialize": initialize,
+        "some_event": some_event
+    }
+
+    ManagerBase = type("ManagerBase", (base_client, ), {
+            "initialize": initialize,
+            "some_event": some_event
+        })
+
+    return ManagerBase