+++ roboticeclient/control/v1/managers/base.py
@@ -13,14 +13,17 @@ TOKEN_FORMAT = "  Token {0}"
 from roboticeclient.common.manager import BaseManager
 
 
-class ControlManager(BaseManager):
+class BaseManager(BaseManager):
 
     api_prefix = '/api'  # /api/v1 etc
 
     def __init__(self, **kwargs):
 
-        self.host = kwargs.pop("host", os.getenv('ROBOTICE_HOST', '127.0.0.1'))
-        self.port = kwargs.pop("port", os.getenv('ROBOTICE_PORT', 9753))
-        self.protocol = kwargs.pop(
-            "protocol", os.getenv('ROBOTICE_PROTOCOL', "http"))
+        if not hasattr(self, "host"):
+            self.host = kwargs.pop("host", os.getenv('ROBOTICE_HOST', '127.0.0.1'))
+        if not hasattr(self, "port"):
+            self.port = kwargs.pop("port", os.getenv('ROBOTICE_PORT', 9753))
+        if not hasattr(self, "protocol"):
+            self.protocol = kwargs.pop(
+                "protocol", os.getenv('ROBOTICE_PROTOCOL', "http"))
         self.set_api()