+++ roboticeclient/control/v1/managers/const.py
@@ -1,0 +1,16 @@
+
+"""
+"""
+
+import sys
+import six
+import logging
+
+from . import base
+
+LOG = logging.getLogger("common")
+
+
+class ConstManager(base.BaseManager):
+
+    SCOPE = "common"